requirements.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
importlib-metadata==3.7.0
Markdown==3.3.4
mysqlclient==2.0.3
Pillow==8.1.2
PyJWT==1.7.1
python-dotenv==0.15.0
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
zipp==3.4.0
